refactor : 닉네임 모달 기능 분리 (score 저장 로직 분리)
