+++ components/modal/NickNameModal.tsx
@@ -5,37 +5,16 @@ import { useSelector, useDispatch } from 'react-redux';
 import { saveSolveUserNameAction } from 'store/user_solve';
 import { RootState } from 'store';
 import { ModalProps } from 'hooks/useModal';
-import SmallOutlinedBtn from 'components/style/button/SmallOutlinedBtn';
-import SmallContainedBtn from 'components/style/button/SmallContainedBtn';
-import { SaveScoreApi } from 'pages/api/quiz';
-import { useRouter } from 'next/router';
-import { saveEmotionCount } from 'store/emotion';
+import { SmallContainedBtn, SmallOutlinedBtn } from 'components/style/button';
 
 const NicknameModal = ({ closeModal }: { closeModal?: ModalProps['closeModal'] }) => {
-  const router = useRouter();
   const dispatch = useDispatch();
-  const { isLoggedin, nickName, userId } = useSelector((state: RootState) => state.user);
+  const { isLoggedin, nickName } = useSelector((state: RootState) => state.user);
   const [_nickname, _nicknameSetter, , _nicknameHandler] = useInput<string>('');
-  const { quizSetId, quizList } = useSelector((state: RootState) => state.solve);
-  const { solveUserScore } = useSelector((state: RootState) => state.user_solve);
 
-  const saveUserData = async () => {
-    try {
-      dispatch(saveSolveUserNameAction({ solveUserName: _nickname }));
-      const res = await SaveScoreApi(
-        _nickname,
-        solveUserScore,
-        quizSetId,
-        isLoggedin ? userId : undefined,
-        quizList.length,
-      );
-      const { quizset_emotion, solver_id } = res.data;
-      dispatch(saveEmotionCount({ quizSetEmotion: quizset_emotion }));
-      router.push(`/quiz/solve/${quizSetId}/result/${solver_id}`);
-      if (closeModal) closeModal();
-    } catch (err) {
-      console.log(err);
-    }
+  const savaSolveUserName = () => {
+    dispatch(saveSolveUserNameAction({ solveUserName: _nickname }));
+    if (closeModal) closeModal();
   };
 
   useEffect(() => {
@@ -48,7 +27,7 @@ const NicknameModal = ({ closeModal }: { closeModal?: ModalProps['closeModal'] }
       <Input type="text" value={_nickname} onChange={_nicknameHandler} placeholder="최대 12자까지" maxLength={12} />
       <ButtonBox>
         <SmallOutlinedBtn onClick={closeModal}>취소</SmallOutlinedBtn>
-        <SmallContainedBtn disabled={!_nickname} onClick={saveUserData}>
+        <SmallContainedBtn disabled={!_nickname} onClick={savaSolveUserName}>
           저장
         </SmallContainedBtn>
       </ButtonBox>
