+++ components/partials/solve/mark/index.ts
@@ -1,5 +1,4 @@
 export { default as MarkList } from './MarkList';
 export { default as MarkCard } from './MarkCard';
-export { default as WrongCard } from './WrongCard';
 export { default as MarkChoiceText } from './MarkChoiceText';
 export { default as MarkChoiceImage } from './MarkChoiceImage';
