+++ components/style/button/index.ts
@@ -1,3 +1,5 @@
 export { default as FlatBtn } from './FlatBtn';
 export { default as LargeContainedBtn } from './LargeContainedBtn';
+export { default as SmallContainedBtn } from './SmallContainedBtn';
 export { default as LargeOutlinedBtn } from './LargeOutlinedBtn';
+export { default as SmallOutlinedBtn } from './SmallOutlinedBtn';
